refactor(analytics-docs): Refactor App.ts
